// src/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
      ingredient_quantity: {
        Row: {
          ingredient_id: number
          quantity: string
          recipe_id: number
        }
        Insert: {
          ingredient_id: number
          quantity: string
          recipe_id: number
        }
        Update: {
          ingredient_id?: number
          quantity?: string
          recipe_id?: number
        }
        Relationships: [
          {
            foreignKeyName: "ingredient_quantity_ingredient_id_fkey"
            columns: ["ingredient_id"]
            isOneToOne: false
            referencedRelation: "ingredients"
            referencedColumns: ["ingredient_id"]
          },
          {
            foreignKeyName: "ingredient_quantity_recipe_id_fkey"
            columns: ["recipe_id"]
            isOneToOne: false
            referencedRelation: "recipes"
            referencedColumns: ["id"]
          },
        ]
      }
      ingredients: {
        Row: {
          ingredient_id: number
          ingredient_name: string
        }
        Insert: {
          ingredient_id?: number
          ingredient_name: string
        }
        Update: {
          ingredient_id?: number
          ingredient_name?: string
        }
        Relationships: []
      }
      instructions: {
        Row: {
          instruction: string
          instruction_id: number
          recipe_id: number
          step_num: number
        }
        Insert: {
          instruction: string
          instruction_id?: number
          recipe_id: number
          step_num: number
        }
        Update: {
          instruction?: string
          instruction_id?: number
          recipe_id?: number
          step_num?: number
        }
        Relationships: [
          {
            foreignKeyName: "instructions_recipe_id_fkey"
            columns: ["recipe_id"]
            isOneToOne: false
            referencedRelation: "recipes"
            referencedColumns: ["id"]
          },
        ]
      }
      recipes: {
        Row: {
          created_at: string
          description: string
          id: number
          image_link: string
          microwave_required: boolean
          original_link: string
          oven_required: boolean
          recipe_name: string
          servings: number
          stove_required: boolean
          total_time_min: number
        }
        Insert: {
          created_at?: string
          description?: string
          id?: number
          image_link?: string
          microwave_required: boolean
          original_link: string
          oven_required: boolean
          recipe_name?: string
          servings: number
          stove_required: boolean
          total_time_min: number
        }
        Update: {
          created_at?: string
          description?: string
          id?: number
          image_link?: string
          microwave_required?: boolean
          original_link?: string
          oven_required?: boolean
          recipe_name?: string
          servings?: number
          stove_required?: boolean
          total_time_min?: number
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never